Moving pdos functions to the dos module, making all spectral functions importable from crawfish.funcs
