+++ src/crawfish/funcs/__init__.py
@@ -1,3 +1,6 @@
-from crawfish.funcs.pcohp import get_pcohp, get_ipcohp
-from crawfish.funcs.pdos import get_pdos, get_ipdos
-from crawfish.funcs.dos import get_dos, get_idos
+from crawfish.funcs.pcohp import (
+    get_pcohp, get_ipcohp, get_pcobi, get_ipcobi, get_pcoop, get_ipcoop, get_pcomo, get_ipcomo
+)
+from crawfish.funcs.dos import (
+    get_dos, get_idos, get_pdos, get_ipdos
+)
